revert. its incompatible with wolfssl at that time. we will focus on makeing a port for it. but for now we have unfortnatly to say on this older openvpn version: add new files
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@65114 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/openvpn/tests/unit_tests/openvpn/mock_get_random.c
@@ -5,7 +5,7 @@
  *             packet encryption, packet authentication, and
  *             packet compression.
  *
- *  Copyright (C) 2017-2026 Sentyron B.V. <openvpn@sentyron.com>
+ *  Copyright (C) 2017-2021 Fox Crypto B.V. <openvpn@foxcrypto.com>
  *
  *  This program is free software; you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License version 2
@@ -17,7 +17,8 @@
  *  GNU General Public License for more details.
  *
  *  You should have received a copy of the GNU General Public License along
- *  with this program; if not, see <https://www.gnu.org/licenses/>.
+ *  with this program; if not, write to the Free Software Foundation, Inc.,
+ *  51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
 #include <stdarg.h>
@@ -40,6 +41,6 @@ prng_bytes(uint8_t *output, int len)
 {
     for (int i = 0; i < len; i++)
     {
-        output[i] = (uint8_t)rand();
+        output[i] = rand();
     }
 }
